+++ todopro-frontend/src/components/__tests__/todo/TodosComponent.spec.js
@@ -1,9 +1,8 @@
 import { mount, flushPromises } from '@vue/test-utils'
-import { describe, it, test, expect, beforeEach, vi } from 'vitest'
+import { describe, test, expect, beforeEach, vi } from 'vitest'
 import TodosComponent from '@/components/todo/TodosComponent.vue'
 import TodoComponent from '@/components/todo/TodoComponent.vue'
 import SideBarComponent from '@/components/partials/SideBarComponent.vue'
-import { router } from '../../../mock-router-setup.js'
 
 const getTodosMock = vi.fn()
 const toastErrorMock = vi.fn()
@@ -54,31 +53,31 @@ describe('TodoListComponent', () => {
     expect(toastErrorMock).toHaveBeenCalledWith('Failed to Fetch Todos on Mount')
   })
 
-  //   test('should go to previous page when prev button is clicked', async () => {
-  //     const currentPage = 4
-  //     wrapper.vm.todosData.currentPage = currentPage
+  test('should go to previous page when prev button is clicked', async () => {
+    const currentPage = 4
+    wrapper.vm.todos.currentPage = currentPage
 
-  //     await wrapper.find('button[aria-label="Previous Page"]').trigger('click')
+    await wrapper.find('button[aria-label="Previous Page"]').trigger('click')
 
-  //     expect(wrapper.vm.todosData.currentPage).toBe(currentPage - 1)
-  //     expect(getTodosMock).toHaveBeenCalledWith(currentPage - 1, expect.anything())
-  //   })
+    expect(wrapper.vm.todos.currentPage).toBe(currentPage - 1)
+    expect(getTodosMock).toHaveBeenCalledWith(currentPage - 1, expect.anything())
+  })
 
   test('should go to next page when next button is clicked', async () => {
     const currentPage = 1
-    wrapper.vm.todosData.currentPage = currentPage
-    wrapper.vm.todosData.totalPages = 3
+    wrapper.vm.todos.currentPage = currentPage
+    wrapper.vm.todos.totalPages = 3
 
     await wrapper.find('button[aria-label="Next Page"]').trigger('click')
 
-    expect(wrapper.vm.todosData.currentPage).toBe(currentPage + 1)
+    expect(wrapper.vm.todos.currentPage).toBe(currentPage + 1)
     expect(getTodosMock).toHaveBeenCalledWith(currentPage + 1, expect.anything())
   })
 
   test('should handle todo deletion', async () => {
     const deletedTodoId = 1
     const currentPage = 1
-    wrapper.vm.todosData.currentPage = currentPage
+    wrapper.vm.todos.currentPage = currentPage
 
     wrapper.vm.handleTodoDeleted(deletedTodoId)
